+++ test/test_cli.rb
@@ -3,10 +3,14 @@
 require 'minitest/autorun'
 require 'minitest/skip_dsl'
 require 'tls_map'
-require 'tls_map//cli/cli'
+require 'tls_map/cli/cli'
 
 class TLSmapCLITest < Minitest::Test
   def test_CLI
     assert(TLSmap::CLI.new)
+  end
+
+  def test_CLI_Extended
+    assert(TLSmap::CLI::Extended.new)
   end
 end